Annex A Template
This is a template for Annex A.  It includes a Requirements Trace Matrix and descriptions of each test.  A test consists of:
The purpose of the test
Pre-conditions for the test (such as the present of necessary OpenAPI elements)
Test Method (what activities should be performed)
References (usually a reference back to the requirements)

The tests in section A.1 assume an OpenAPI landing page.  Additional sections can be added for other landing page formats