Update mapping between spec URIs and spec elements
I did the following changes:

* several spec URIs had multiple mappings, I have deleted all the incorrect ones
* the spec URIs for the conformance classes have been missing and have been added
* I noticed four test cases which are not included in the current spec document (two are mentioned, but not included), I have added them and added links to the test case in the GitHub repository

I also noticed the following other issues:

* The test cases use `.../ats/{confClassId}/{testCaseId}` while the conformance classes use `.../conf/{confClassId}`. I checked the rules for specification elements and noticed that we should have used `.../conf/{confClassId}/{testCaseId}` for the test cases. 
* For the OAS 3.0 conformance class we use `oas3` for the id, but everywhere else we use `oas30` (requirements classes, normative statements, test cases). 

My proposal is to fix all these issues in 1.0.1 (whenever we will work on it) and continue to support the then deprecated URIs (e.g. the `ats` and `oas3` URIs). We would then also redirect the URIs of the four missing test cases when the test cases have anchors in the document.