Clarify wording on spatial literals
As discussed in the SWG call today, the `basic-spatial-functions` requirements class does not restrict the spatial data types that may be used in a property, it only restricts the literal values. The `basic-spatial-functions-plus` requirements class supports more literal values. It would be clearer to use "spatial literals" in some cases where currently "spatial data types" is used.